contd.
