--- BasicRenderer/src/Resources/SubresourceView.cpp
@@ -1,3 +1,0 @@
-#include "Resources/SubresourceView.h"
-
-#include "Resources/Resource.h"
